Remove more warnings

+++ test/bfp_tests/src/bfp/real/test_bfp_convolve.c
@@ -61,29 +61,29 @@ TEST(bfp_convolve, vect_s32_convolve_valid)
     const unsigned P = tap_count >> 1;
 
     right_shift_t shr = pseudo_rand_uint(&seed, 0, 6);
-    for(int k = 0; k < length; k++)
+    for(unsigned int k = 0; k < length; k++)
       signal_in[k] = pseudo_rand_int32(&seed) >> shr;
 
     bfp_s32_init(&bfp_in, signal_in, pseudo_rand_int(&seed, -20, 20), length, 1);
     bfp_s32_init(&bfp_out, signal_out, pseudo_rand_int(&seed, -20, 20), length-2*P, 0);
 
-    for(int k = 0; k < tap_count; k++)
+     for(unsigned int k = 0; k < tap_count; k++)
       taps[k] = pseudo_rand_uint32(&seed) >> 1;
 
     int64_t tap_total = vect_s32_sum(taps, tap_count);
 
     float scale = ldexpf(1, 30) / tap_total;
 
-    for(int k = 0; k < tap_count; k++)
-      taps[k] *= scale;
+     for(unsigned int k = 0; k < tap_count; k++)
+      taps[k] *= (int32_t) scale;
 
     const unsigned out_length = length - ((tap_count>>1)<<1);
 
-    for(int k = 0; k < out_length; k++){
+    for(unsigned int k = 0; k < out_length; k++){
       vpu_int32_acc_t acc = 0;
-      for(int p = 0; p < tap_count; p++)
+      for(unsigned int p = 0; p < tap_count; p++)
         acc = vlmacc32(acc, signal_in[k+p], taps[p]);
-      expected[k] = acc;
+      expected[k] = (int32_t) acc;
     }
     
     bfp_s32_convolve_valid(&bfp_out, &bfp_in, taps, tap_count);
@@ -123,30 +123,30 @@ TEST(bfp_convolve, vect_s32_convolve_same_reflected)
     const int P = tap_count >> 1;
 
     right_shift_t shr = pseudo_rand_uint(&seed, 0, 6);
-    for(int k = 0; k < length; k++)
+    for(unsigned int k = 0; k < length; k++)
       signal_in[k] = pseudo_rand_int32(&seed) >> shr;
 
     bfp_s32_init(&bfp_in, signal_in, pseudo_rand_int(&seed, -20, 20), length, 1);
     bfp_s32_init(&bfp_out, signal_out, pseudo_rand_int(&seed, -20, 20), length, 0);
 
-    for(int k = 0; k < tap_count; k++)
+     for(unsigned int k = 0; k < tap_count; k++)
       taps[k] = pseudo_rand_uint32(&seed) >> 1;
 
     int64_t tap_total = vect_s32_sum(taps, tap_count);
 
     float scale = ldexpf(1, 30) / tap_total;
 
-    for(int k = 0; k < tap_count; k++)
-      taps[k] *= scale;
+     for(unsigned int k = 0; k < tap_count; k++)
+      taps[k] *= (int32_t) scale;
 
-    for(int k = 0; k < length; k++){
+    for(unsigned int k = 0; k < length; k++){
       vpu_int32_acc_t acc = 0;
-      for(int t = 0; t < tap_count; t++) {
+      for(unsigned int t = 0; t < tap_count; t++) {
         int i = k + t - P;
-        int32_t input = signal_in[ (i < 0)? (-i) : (i >= length)? (2*length - i - 2) : i ];
+        int32_t input = signal_in[ (i < 0)? (-i) : (i >= (int) length)? (2*length - i - 2) : i ];
         acc = vlmacc32(acc, input, taps[t]);
       }
-      expected[k] = acc;
+      expected[k] = (int32_t) acc;
     }
 
     memset(signal_out, 0, sizeof(signal_out));
@@ -191,32 +191,32 @@ TEST(bfp_convolve, vect_s32_convolve_same_zero)
     const int P = tap_count >> 1;
 
     right_shift_t shr = pseudo_rand_uint(&seed, 0, 6);
-    for(int k = 0; k < length; k++)
+    for(unsigned int k = 0; k < length; k++)
       signal_in[k] = pseudo_rand_int32(&seed) >> shr;
 
     bfp_s32_init(&bfp_in, signal_in, pseudo_rand_int(&seed, -20, 20), length, 1);
     bfp_s32_init(&bfp_out, signal_out, pseudo_rand_int(&seed, -20, 20), length, 0);
 
-    for(int k = 0; k < tap_count; k++)
+     for(unsigned int k = 0; k < tap_count; k++)
       taps[k] = pseudo_rand_uint32(&seed) >> 1;
 
     int64_t tap_total = vect_s32_sum(taps, tap_count);
 
     float scale = ldexpf(1, 30) / tap_total;
 
-    for(int k = 0; k < tap_count; k++)
-      taps[k] *= scale;
+     for(unsigned int k = 0; k < tap_count; k++)
+      taps[k] *= (int32_t) scale;
 
-    for(int k = 0; k < length; k++){
+    for(unsigned int k = 0; k < length; k++){
       vpu_int32_acc_t acc = 0;
-      for(int t = 0; t < tap_count; t++) {
+      for(unsigned int t = 0; t < tap_count; t++) {
         int i = k + t - P;
         int32_t input = signal_in[ i ];
-        if( (i < 0) || (i >= length) )
+        if( (i < 0) || (i >= (int) length) )
           input = 0;
         acc = vlmacc32(acc, input, taps[t]);
       }
-      expected[k] = acc;
+      expected[k] = (int32_t) acc;
     }
 
     memset(signal_out, 0, sizeof(signal_out));
@@ -261,32 +261,32 @@ TEST(bfp_convolve, vect_s32_convolve_same_extend)
     const int P = tap_count >> 1;
 
     right_shift_t shr = pseudo_rand_uint(&seed, 0, 6);
-    for(int k = 0; k < length; k++)
+    for(unsigned int k = 0; k < length; k++)
       signal_in[k] = pseudo_rand_int32(&seed) >> shr;
 
     bfp_s32_init(&bfp_in, signal_in, pseudo_rand_int(&seed, -20, 20), length, 1);
     bfp_s32_init(&bfp_out, signal_out, pseudo_rand_int(&seed, -20, 20), length, 0);
 
-    for(int k = 0; k < tap_count; k++)
+     for(unsigned int k = 0; k < tap_count; k++)
       taps[k] = pseudo_rand_uint32(&seed) >> 1;
 
     int64_t tap_total = vect_s32_sum(taps, tap_count);
 
     float scale = ldexpf(1, 30) / tap_total;
 
-    for(int k = 0; k < tap_count; k++)
-      taps[k] *= scale;
+     for(unsigned int k = 0; k < tap_count; k++)
+      taps[k] *= (int32_t) scale;
 
-    for(int k = 0; k < length; k++){
+    for(unsigned int k = 0; k < length; k++){
       vpu_int32_acc_t acc = 0;
-      for(int t = 0; t < tap_count; t++) {
+      for(unsigned int t = 0; t < tap_count; t++) {
         int i = k + t - P;
         int32_t input = signal_in[ i ];
         if( i < 0 )             input = signal_in[0];
-        else if( i >= length )  input = signal_in[length - 1];
+        else if( i >= (int) length )  input = signal_in[length - 1];
         acc = vlmacc32(acc, input, taps[t]);
       }
-      expected[k] = acc;
+      expected[k] = (int32_t) acc;
     }
 
     memset(signal_out, 0, sizeof(signal_out));
